% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centrimpact-package.R
\docType{package}
\name{centrimpact-package}
\alias{centrimpact}
\alias{centrimpact-package}
\title{centrimpact: Supporting the Analysis and Visualization of Community Engaged Research Metrics}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides tools for analyzing and visualizing community-engaged research metrics within the CEnTR-IMPACT framework. The package includes functions for network analysis, diversity calculations, proportional scaling, and creating consistent visualizations with customizable color palettes and typography. Designed to support researchers in evaluating community engagement outcomes and impact across multiple domains and network layers.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/CENTRIMPACT/centrimpact}
  \item Report bugs at \url{https://github.com/CENTRIMPACT/centrimpact/issues}
}

}
\author{
\strong{Maintainer}: Jeremy Price \email{jfprice@iu.edu} (\href{https://orcid.org/0000-0002-6506-3526}{ORCID})

}
\keyword{internal}
